Add Play,Add,Enqueue actions to track
The play command has been revamped to play the first instance of a song
in the playlist, and to add it if it doesn't exist.
Add just has been given an action. It appends to playlist.
Enqueue adds the song right after the current playing song. Hopefully
later it will monitor queued tracks and remove them after playing, à la
winamp, xmms.